requirements.txt
selenium==2.52.0
numpy==1.10.4 
pillow==3.1.1
https://storage.googleapis.com/tensorflow/mac/tensorflow-0.7.0-py2-none-any.whl